Update GuildAccountsModule.cs
C#'s casting can suck my dick
Removed snarky comments

+++ Miki/Modules/GuildAccountsModule.cs
@@ -274,10 +274,7 @@ namespace Miki.Modules
 
             using (var context = new MikiContext())
             {
-                int totalGuilds = Ceiling(await context.GuildUsers.CountAsync() / amountToTake);
-                // int totalGuilds = await context.GuildUsers.CountAsync() / 12;
-                // This should ceiling, since when this is corrected going to the last page will be off by 1
-                // Also veld why the fuck was this hardcoded.
+                int totalGuilds = (int) Ceiling((double) await context.GuildUsers.CountAsync() / amountToTake);
 
                 List<GuildUser> leaderboards = await context.GuildUsers.OrderByDescending(x => x.Experience)
                                                                       .Skip(amountToSkip * amountToTake)
